split personal + confirmation questions u, questions 1-13
