+++ .vscode/settings.json
@@ -10,5 +10,8 @@
 	"mypy-type-checker.cwd": "${workspaceFolder}/apps/api",
 	"flake8.cwd": "${workspaceFolder}/apps/api",
 	"isort.args": ["--settings-path", "apps/api/pyproject.toml"],
-	"python.testing.pytestArgs": ["--rootdir", "apps/api"]
+	"python.testing.pytestArgs": [
+		"--rootdir",
+		"apps/api"
+	]
 }
