+++ apps/site/src/lib/components/forms/DropdownSelect.tsx
@@ -19,7 +19,7 @@ interface OtherProps {
 const OtherPopup = ({ value, name }: OtherProps) => {
 	if (value === "other") {
 		return (
-			<div className="mt-2 flex gap-2">
+			<div className="mt-2 flex flex-row gap-2">
 				<label htmlFor={`${name}-other-input`} className="text-lg">
 					Other: <RequiredAsterisk />
 				</label>
@@ -27,7 +27,7 @@ const OtherPopup = ({ value, name }: OtherProps) => {
 					type="text"
 					name={`_other_${name}`}
 					id={`${name}-other-input`}
-					className="text-black border-b-2 p-1 h-6 border-black w-6/12"
+					className="text-[var(--color-black)] text-sm border shadow-lg p-2 h-10 resize-none rounded-xl flex flex-col"
 					required
 				/>
 			</div>
@@ -45,26 +45,28 @@ export default function DropdownSelect({
 
 	return (
 		<div className={containerClass}>
-			<label className="text-lg mb-2" htmlFor={name}>
-				{labelText} <RequiredAsterisk />
-			</label>
-			<select
-				className="bg-[#e1e1e1] text-[var(--color-black)] text-lg h-10 p-1.5 rounded-md"
-				name={name}
-				id={name}
-				defaultValue={""}
-				onChange={(e) => setValue(e.target.value)}
-				required
-			>
-				<option value="" disabled />
-				{values.map((item, i) => {
-					return (
-						<option key={`option-${i}`} value={item.value}>
-							{item.text}
-						</option>
-					);
-				})}
-			</select>
+			<div className="flex flex-col">
+				<label className="text-md mb-2" htmlFor={name}>
+					{labelText} <RequiredAsterisk />
+				</label>
+				<select
+					className=" text-[var(--color-black)] text-lg h-10 p-1.5 rounded-xl border shadow-lg"
+					name={name}
+					id={name}
+					defaultValue={""}
+					onChange={(e) => setValue(e.target.value)}
+					required
+				>
+					<option value="" disabled />
+					{values.map((item, i) => {
+						return (
+							<option key={`option-${i}`} value={item.value}>
+								{item.text}
+							</option>
+						);
+					})}
+				</select>
+			</div>
 			<OtherPopup value={value} name={name} />
 		</div>
 	);
